변경사항 1. int->Long 으로 전부 수정 2. 비어있는 엔드포인트에서 () 삭제 3. 일부 메소드에 남아있던 "/series" 삭제 4. alarm 컨트롤러 메소드의 엔드포인트에 {userId} cnrk

+++ src/main/java/com/hunnit_beasts/kelog/controller/EtcController.java
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.RestController;
 public class EtcController {
 
     @GetMapping("/view-cnt/{postId}")
-    public void viewCount(@PathVariable(value = "postId") int postId) {}
+    public void viewCount(@PathVariable(value = "postId") Long postId) {}
 }
